refactored population classes
